extract MoneyRequestReceiptView
(cherry picked from commit 887bc1111dd5f1cd5f84d9396a1de74a14c72e1b)
